+++ frontend/src/scenes/experiments/charts/funnel/funnelUtils.ts
